rename [윤아영] : 타입 파일 이름 변경, import시 타입임을 명시
