+++ components/chips/tag-chip.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren } from 'react';
+import type { Color } from '@/components/chips/color.type';
 import { tagColorConfig } from '@/components/chips/color-constants';
-import { Color } from '@/components/chips/color-type';
 
 const TagChip = ({ children, color }: PropsWithChildren<{ color: Color }>) => {
   return (
